+++ svd-rs/src/field.rs
@@ -1,89 +1,54 @@
-use super::{DimElement, FieldInfo};
 use core::ops::{Deref, DerefMut};
 
-/// Describes a field or fields of a [register](crate::RegisterInfo).
+use super::{DimElement, FieldInfo};
+
+/// A single field or array of fields
+#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
 #[derive(Clone, Debug, PartialEq)]
-pub enum Field {
-    /// A single field.
-    Single(FieldInfo),
-    /// A field array.
-    Array(FieldInfo, DimElement),
+pub struct Field {
+    #[cfg_attr(feature = "serde", serde(flatten))]
+    /// A description of a field
+    pub info: FieldInfo,
+    #[cfg_attr(
+        feature = "serde",
+        serde(flatten, default, skip_serializing_if = "Option::is_none")
+    )]
+    /// If `None` it is a single field, if `Some` specifies array attributes
+    pub dim: Option<DimElement>,
 }
 
 impl Deref for Field {
     type Target = FieldInfo;
 
     fn deref(&self) -> &FieldInfo {
-        match self {
-            Self::Single(info) => info,
-            Self::Array(info, _) => info,
-        }
+        &self.info
     }
 }
 
 impl DerefMut for Field {
     fn deref_mut(&mut self) -> &mut FieldInfo {
-        match self {
-            Self::Single(info) => info,
-            Self::Array(info, _) => info,
-        }
+        &mut self.info
     }
 }
 
 impl Field {
+    /// Construct single [`Field`]
+    pub const fn single(info: FieldInfo) -> Self {
+        Self { info, dim: None }
+    }
+    /// Construct [`Field`] array
+    pub const fn array(info: FieldInfo, dim: DimElement) -> Self {
+        Self {
+            info,
+            dim: Some(dim),
+        }
+    }
     /// Return `true` if field instance is single
     pub const fn is_single(&self) -> bool {
-        matches!(self, Self::Single(_))
+        matches!(&self.dim, None)
     }
     /// Return `true` if it is field array
     pub const fn is_array(&self) -> bool {
-        matches!(self, Self::Array(_, _))
-    }
-}
-
-#[cfg(feature = "serde")]
-mod ser_de {
-    use super::*;
-    use serde::{Deserialize, Deserializer, Serialize, Serializer};
-
-    #[derive(serde::Deserialize, serde::Serialize)]
-    struct FieldArray {
-        #[cfg_attr(
-            feature = "serde",
-            serde(flatten, default, skip_serializing_if = "Option::is_none")
-        )]
-        dim: Option<DimElement>,
-        #[cfg_attr(feature = "serde", serde(flatten))]
-        info: FieldInfo,
-    }
-
-    impl Serialize for Field {
-        fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-        where
-            S: Serializer,
-        {
-            match self {
-                Self::Single(info) => info.serialize(serializer),
-                Self::Array(info, dim) => FieldArray {
-                    dim: Some(dim.clone()),
-                    info: info.clone(),
-                }
-                .serialize(serializer),
-            }
-        }
-    }
-
-    impl<'de> Deserialize<'de> for Field {
-        fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-        where
-            D: Deserializer<'de>,
-        {
-            let FieldArray { dim, info } = FieldArray::deserialize(deserializer)?;
-            if let Some(dim) = dim {
-                Ok(info.array(dim))
-            } else {
-                Ok(info.single())
-            }
-        }
+        matches!(&self.dim, Some(_))
     }
 }