+++ svd-rs/src/register.rs
@@ -2,43 +2,54 @@ use core::ops::{Deref, DerefMut};
 
 use super::{DimElement, RegisterCluster, RegisterInfo};
 
-/// A single register or array of registers. A register is a named, programmable resource that belongs to a [peripheral](crate::Peripheral).
+/// A single register or array of registers
+#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
 #[derive(Clone, Debug, PartialEq)]
-pub enum Register {
-    /// A single register.
-    Single(RegisterInfo),
-    /// An array of registers.
-    Array(RegisterInfo, DimElement),
+pub struct Register {
+    #[cfg_attr(feature = "serde", serde(flatten))]
+    /// A description of a register
+    pub info: RegisterInfo,
+    #[cfg_attr(
+        feature = "serde",
+        serde(flatten, default, skip_serializing_if = "Option::is_none")
+    )]
+    /// If `None` it is a single register, if `Some` specifies array attributes
+    pub dim: Option<DimElement>,
 }
 
 impl Deref for Register {
     type Target = RegisterInfo;
 
     fn deref(&self) -> &RegisterInfo {
-        match self {
-            Self::Single(info) => info,
-            Self::Array(info, _) => info,
-        }
+        &self.info
     }
 }
 
 impl DerefMut for Register {
     fn deref_mut(&mut self) -> &mut RegisterInfo {
-        match self {
-            Self::Single(info) => info,
-            Self::Array(info, _) => info,
-        }
+        &mut self.info
     }
 }
 
 impl Register {
+    /// Construct single [`Register`]
+    pub const fn single(info: RegisterInfo) -> Self {
+        Self { info, dim: None }
+    }
+    /// Construct [`Register`] array
+    pub const fn array(info: RegisterInfo, dim: DimElement) -> Self {
+        Self {
+            info,
+            dim: Some(dim),
+        }
+    }
     /// Return `true` if register instance is single
     pub const fn is_single(&self) -> bool {
-        matches!(self, Self::Single(_))
+        matches!(&self.dim, None)
     }
     /// Return `true` if it is register array
     pub const fn is_array(&self) -> bool {
-        matches!(self, Self::Array(_, _))
+        matches!(&self.dim, Some(_))
     }
 }
 
@@ -87,52 +98,5 @@ impl<'a> std::iter::Iterator for RegIterMut<'a> {
             }
         }
         None
-    }
-}
-
-#[cfg(feature = "serde")]
-mod ser_de {
-    use super::*;
-    use serde::{Deserialize, Deserializer, Serialize, Serializer};
-
-    #[derive(serde::Deserialize, serde::Serialize)]
-    struct RegisterArray {
-        #[cfg_attr(
-            feature = "serde",
-            serde(flatten, default, skip_serializing_if = "Option::is_none")
-        )]
-        dim: Option<DimElement>,
-        #[cfg_attr(feature = "serde", serde(flatten))]
-        info: RegisterInfo,
-    }
-
-    impl Serialize for Register {
-        fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-        where
-            S: Serializer,
-        {
-            match self {
-                Self::Single(info) => info.serialize(serializer),
-                Self::Array(info, dim) => RegisterArray {
-                    dim: Some(dim.clone()),
-                    info: info.clone(),
-                }
-                .serialize(serializer),
-            }
-        }
-    }
-
-    impl<'de> Deserialize<'de> for Register {
-        fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-        where
-            D: Deserializer<'de>,
-        {
-            let RegisterArray { dim, info } = RegisterArray::deserialize(deserializer)?;
-            if let Some(dim) = dim {
-                Ok(info.array(dim))
-            } else {
-                Ok(info.single())
-            }
-        }
     }
 }